requirements.txt
azure-cognitiveservices-speech
streamlit